requirements.txt
fastapi
uvicorn[standard]
teradatasql
python-dotenv
